core: handle host architecture in fence flow
This patch adds a getArchitecture method to VDS.
It is used in order to get the host architecture using its cpu flags, in
case that this is null, the architecture of the containing cluster is
used.

Please note that a separate call to get VdsDynamic information for the
host since the VDS object passed to the fencing code is slim and does
not contain any dymanic information.

Change-Id: I081361c9f956ea71a715acf0368c79c7130674c7
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1149235
Signed-off-by: Eli Mesika <emesika@redhat.com>