core: assume live snapshot supported as default
VDSM will report if underlying QEMU reports
live snapshot or not. This information is
provided through libvirt.

If libvirt, or VDSM, no longer report
this capability, Engine get stuck with
the old value and it is no longer updated.

There is an unlucky case on CentOS and RHEL 6.6
on which we can hit this condition.

To fix that, this patch assumes live snapshot
is supported if we don't have word from VDSM.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1159211
Related-To: http://bugzilla.redhat.com/1149460
Related-To: http://bugzilla.redhat.com/1158974
Change-Id: I4660560e08f606e2b16a580ef60618abfad49244
Signed-off-by: Francesco Romani <fromani@redhat.com>