core: free pending resources properly on migration failure
There is a race when VM switch from MIGRATING_FROM to UP due to
migration failure between:
1. Rerun operation which is executed from the monitoring (VURTI) on a
separate thread
2. Powering up handling which is executed from the monitoring thread

When #2 happens before #1, the powering up handling decrease the pending
resources from the destination host properly.

But when #1 happens before, then we've got a problem because it sets the
destination host to null so neither #1 or #2 decreases the pending
resources on the destination host.

This patch solves the pending resources decrement in the rerun flow.

Change-Id: Ib866d7da8a3e1fa7fe2023adc49f3b4f559c8982
Bug-Url: https://bugzilla.redhat.com/1157211
Signed-off-by: Arik Hadas <ahadas@redhat.com>