core: New tests for AddVmTemplateCommand
While making a new template based on a VM, providing the template's
allocated disks' alias as an empty string is now disabled. The user
should not be able to create a template with disks that have an empty
alias. This patch adds tests to AddVmTemplateCommandTest to verify
that this will be disabled at the backend level.

Change-Id: I614239c750096b1c3585e0e56bc0d77d328e4c8f
Bug-Url: https://bugzilla.redhat.com/1110304
Signed-off-by: Amit Aviram <aaviram@redhat.com>