core: Add Video managed device for VM's/Template's OVF.
The number of monitors data in the Template should be part of the OVF
data.
so the information will still be availale after the attach of
a Storage Domain with import of existing entities.

This information consists in the video device of the entity, therefore
the fix adds video devices while building the OVF data, same as being
done in the export VM/Template process.

Change-Id: I5583edadd75bce7dfb3f5ba04cfcbe38f1dc7091
Bug-Url: https://bugzilla.redhat.com/1136902
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>