restapi: Avoid NPE in request for CDROMs of wrong VM
Currently when we receive a request for the collection of CDROMs of a VM
we don't take into account that the query performed to find the VM can
return null if there isn't such a VM (the identifier is wrong, or the VM
has been removed). To avoid that NPE this patch checks the result of the
query and will respond with an empty list in those cases.

Change-Id: Id17ef36193c21e113d4abafeee0c769a2a948bd3
Bug-Url: https://bugzilla.redhat.com/1150953
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>