webadmin: Correction of UI Constants
Messages that was using the word "Detach" incorrectly, namely using
capital "D" when there was no need to- changed to "detach".
In addition, the "Data Center" part of the message who related
to detaching storage from a data center has changed back to
"Data Center(s)" due to the fact that it is possible that some
chosen Storage is attached to more than one Data Center.

Change-Id: Id288103a417db6aee13e3f990b2d0af3fb3827f7
Bug-Url: https://bugzilla.redhat.com/1152558
Relates-To:https://bugzilla.redhat.com/1149457
Signed-off-by: Amit Aviram <aaviram@redhat.com>