webadmin: changed the list of fields which can not be overridden
There are the following types of fields in instance types:
- the ones which are taken from it but the user can override them (e.g. defaults)
- the ones which the user can not override

This patch reduces the num of fields which can not be overridden to contain
only the really most important ones. This avoids the situation that the user will
have to have instance types like: "big with soundcard", "big with smardcard and
soundcard" etc.

The list is now significantly smaller, contains only:
- memory size
- num of sockets/cores per socket
- HA
- migration model/downtime
- priority
- balloon/min allocated memory

The specific UI changes include:
- remove the chain icon from fields which are not considered to be "marked"
  anymore
- remove the specific fields from
  InstanceTypeAttachDetachManager.detachableModels so it will not detach the VM
  from instance type if changed

Change-Id: I7fc831f0a0656f8eaa92fe3509b03c8ac1b71fb9
Bug-Url: https://bugzilla.redhat.com/1149663
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>