core: changed the list of fields which can not be overridden
Some of the fields on the instance types are prohibited to be overridden
because in such a case the VM would not be considered to be based on that
instance type anymore.

This patch reduces the num of fields which can not be overridden to contain
only the really most important ones. This avoids the situation that the user will
have to have instance types like: "big with soundcard", "big with smardcard and
soundcard" etc.

The list is now significantly smaller, contains only:
- memory size
- num of sockets/cores per socket
- HA
- migration model/downtime
- priority
- balloon/min allocated memory

Change-Id: I6142a240d60d54634361d3e6427b2a56fbd01362
Bug-Url: https://bugzilla.redhat.com/1149663
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>