core: change import vm to be VmCommand - part 3
Introduce ImportValidator that stores the validation logics which are
required for import operations, to reduce the coupling between
ImportVmcommand and MoveOrCopyTemplateCommand. This validator will be
used later in other import commands.

The tests which test the logic of the validation were moved to
ImportValidatorTest and the tests in ImportVmCommandTest were changed
such that they will check that the right error is returned from the
canDoAction method, assuming that the validator returns the appropriate
error.

Change-Id: I7fa61e7f493d196acdb40bfe461665d6668c6508
Signed-off-by: Arik Hadas <ahadas@redhat.com>