core: perform ovf update for domain when it's deactivated
When a storage domain is being deactivated, we can attempt to perform an
ovf update to that domain if the data center version supports it, if the
update fails the domain will still be moved to maintenance.

Ovf stores are created for a domain according to the config
value StorageDomainOvfStoreCount on each OvfDataUpdater run, so the same
case should be supported on deactivation of domain (OVF stores will be
created if needed).

Change-Id: I54cc929937d3b14dc49a2edab9e763cbe603a9c3
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1138124
Signed-off-by: Liron Aravot <laravot@redhat.com>