webadmin: Added range values to QoS error message.
On the Data Center tab, When adding a QoS to the cpu with limit which
is out of range. The error message doesn't state what is the legal
limit range.

Now the range is added to the error message(The range is 1-100).
Also, changed the name of the variable:
ACTION_TYPE_FAILED_QOS_OUT_OF_RANGE_VALUES ->
ACTION_TYPE_FAILED_QOS_CPU_OUT_OF_RANGE_VALUES

A new variable for the "out of range" error was added for the storage
which is named:
ACTION_TYPE_FAILED_QOS_STORAGE_OUT_OF_RANGE_VALUES

Change-Id: Ia0e1b2ef97624b6b0f3c9c95e7a09e15561a458b
Bug-Url: https://bugzilla.redhat.com/1142333
Signed-off-by: Tomer Saban <tsaban@redhat.com>