restapi: copy instance type fields to the VM
As a consequence of the https://bugzilla.redhat.com/1149663 the instance type
fields are not copied to the VM on server anymore and frontends
(WA/UP/REST) need to take care of it.

More specifically it works like this:

If only the instance type is selected, it has to copy its properties to the
newly created VM (on VM level this can be overridden)

If both the template and the instance type is selected, the instance type
overrides the template.

When copying the parameters from instance type to the VM, only the ones which
are supported on the target cluster/OS are taken into account.

Change-Id: Id8ff7a39fe74c73d8984fb34169b69649dab7c15
Bug-Url: https://bugzilla.redhat.com/1149663
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>