core:fix handling of admin user while login
GetAllRolesByUserIdAndGroupIds was called whenever a user was logged in
in order to find out if the user is an admin or not.
In original code, all roles were retrieved and the check for the admin
role type was done in a loop inside the
MultiLevelAdministrationHandler::isAdminUser() code.

Since the required functionality is only to find if any admin role is
attached to the logging user, I had changed the SP code to return only
one role (if exists) and have the role_type checked in the SP level.

SP name and DAO method names were changed to match the real
functionality.

Change-Id: I6aa489199c904008e46a650f11877091931ee5de
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1156465
Signed-off-by: Eli Mesika <emesika@redhat.com>