Allow registerVds with same name and hostname if provisioned
Currently we delete the old vds record with InstallingOS status to allow
adding new record as waiting for approval. This is wrong approach
which causes lose of all the initial information that set on the vds record.

With this fix we don't delete the record and skip the duplication check

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1115343
Change-Id: I157b567b2c868b12a89a394e909e97895338380d
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>