Revert "core: Store iSCSI Target Address instead of IP Address in DB"
This reverts commit c6c9bdb239042ca0ab4543d154d5144740ba4379.
The idea of working with the original address we get from the user
instead of the IP address we get from VDSM works fine in case that the
iscsi server has only one NIC. When it has more than one (Multipath, for
example) the VDSM sends all the addresses it discovered, and we override
all of them with the original address we got from the user. This makes
the login to all of the other targets fail (except for the first one).
Thus, working with the original address is not possible.

Change-Id: I3974820e9039f05aa716752330c1b2ec13591137
Signed-off-by: Idan Shaby <ishaby@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1086310