core:enable to filter by search query and user
This patch adds support for user search queries that considers both the
search expression as given by the user and the filtering that should
apply in order to cope with user permissions.

This is done by doing the regular filtering query (SP) followed by the
search query (without filtering) , then the intersection of this two
results is returned.

Change-Id: Ic8beef0abbac999e49a726f1fecd9bcabec2a923
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1163105
Signed-off-by: emesika <emesika@redhat.com>