engine: Enhance the xml-rpc Long support to deal with Maps and Lists
It is common to have a Long value in a sub-map of an xml-rpc
argument. The current implementation does not deal with this case.

The ideal solution to this would be to instruct the marshaller
to do this. Unfortunately that is not possible and we have to
work around that limitation.

Change-Id: I5a5933a68f937c4734b547c173188e88bd7b9464
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1145665
Signed-off-by: Martin Sivák <msivak@redhat.com>