core: check if the HE guest can be migrated before maintenace
When putting VDS to maintenance mode the HE guest is migrated by the
ha-agent so it's not checked by the engine at all. In case there is no
VDS to migrate the HE guest to the migration will never happen and
the host stays in "preparing for maintenance" until the admin activates
it manually. This patch adds the logic to check if there is a VDS
which can be used for HE guest. It's a bit racy, but it will prevent
at least the case described in the referenced bugzilla.

Change-Id: I0b9cbcd03f01ca34c6c31308b7321214e15cafc7
Bug-Url: https://bugzilla.redhat.com/1157378
Signed-off-by: Jiri Moskovcak <jmoskovc@redhat.com>