restapi: Support multiple links to DC from SD
Currently the Storage Domains available in the top level /storagedomains
collection don't have a reference to the Data Centers that they are
attached to. Only the /datacenters/{datacenter:id}/storagedomains
collection provides this information, but in the context of particular
Data Center, so there is no way to easily find out to which Data Centers
a Storage Domain is attached to.  In general a Storage Domain may be
attached to multiple Data Centers simultaneously (an ISO Storage Domain
in particular). To simplify finding the Data Centers this patch adds a
new "data_centers" attribute to all the Storage Domain collections. This
will contain a set of links to the Data Centers. For example:

  GET /storagedomains/{storagedomain:id}
  <storage_domain id="..." href="...">
    <name>mydata</name>
    <data_centers>
      <data_center id="..." href="..."/>
      <data_center id="..." href="..."/>
      ...
    </data_centers>
  </storage_domain>

The previously existing "data_center" (capable of holding one Data
Center only) will be preserved for backwards compatibility, but it will
removed in the next major release.

Change-Id: I55c52309645f73d7a790c07135c21acf91feb824
Bug-Url: https://bugzilla.redhat.com/1132499
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>