foreman-integration: Remove domain from hostname to allow rhevh registration
After RHEVH provisioning installation the provisioned host sends registration
request which includes and name added to the domain url as the vds name.

oVirt adds the host only with the name, therefore we need to remove the
domain url from the request to find out if the request comes after
provision. In such case we will update the vds record as it is without
relying on the request parameters.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1115343
Change-Id: Ie95c258def25d2e484d4094784eb0b15d62c7ef3
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>