core: update vm version for pool-vm
When pool-vm is started by admin, it is not attached to any user, nor
doest it have a stateless snapshot.
so in this case, the logic that check if new version exists is skipped.

this patch calls restoreStatelessSnapshot (that contains the new version
logic) also to pool vms that are not attached to any user.

Change-Id: Ibce7bc6d744a0bc4ef42e7f039e880c595e44f89
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1167102
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>