core: Monitoring - refactor Host newtork error handling
Host VDS network error should trigger a state change
and a treatment for auto-fencing as outlined in [1]

[1] http://www.ovirt.org/Automatic_Fencing

This handling is done at a single place [2] and is only called
from a single place, the VdsEventListener [3].

[2] VdsManager.handleNetworkException

[3] void onError(@Observes VdsNetworkException networkException)

*All* vds broker commands in action that throw VDS network exception is also
firing and event with the Exception as payload.

The event it caught by VdsEventListener and it invokes the handling on
using the assigned VdsManager in a separate thread.

Mo more VDSNetwork Exception throwing and re-throwing for flow control.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1099068
Change-Id: I59d26baa4d372905c8249edb0aabe28ea4c4a74a
Signed-off-by: Roy Golan <rgolan@redhat.com>