core: Avoid too many IDE or sPAPR VSCSI devices
If there are any payload devices or cloud-init is enabled, it is
possible to exceed the number of allowed sPAPR VSCSI (in ppc64) or IDE
(in x86-64) devices by having the virtual CD, the payload device and
a large number of disks in these interfaces.

This patch warns the user if that is the case and prevents the VM from
running.

Change-Id: Ief86ea10237c931490b96affaa2f97e32d617c3b
Bug-Url: https://bugzilla.redhat.com/1076488
Signed-off-by: Shahar Havivi<shavivi@redhat.com>