core: decrement pending vm on 'run in paused mode'
In a flow where VM is started in paused state neither of
runningSucceeded nor runningFailed methods in RunVmCommand are called.
The command is kept in the collection of async commands until the VM
will be resumed.

On resume, the reportCompleted method of the original RunVmCommand will
be called and it will be removed from the async command collection.
Thus, the pending resources are not cleared in this flow.

The solution is to decrease the pending resources in the reportCompleted
method if they were not decreased before.

Change-Id: Icc931118a848c1c7507c407d955537761dd1a742
Bug-Url: https://bugzilla.redhat.com/1174816
Signed-off-by: Arik Hadas <ahadas@redhat.com>