packaging: Use correct Avalon Framework .jar files in EL7
In version 7 of RHEL and CentOS the Avalon Framework .jar files are
split, like in Fedora, but currently we assume that they aren't, as that
was the case in version 6. This modifies the .spec file so that this
difference is taken into account.

Change-Id: Idd0ea12dfbd3ad695cc2a3390fd3d104d2823ea9
Bug-Url: https://bugzilla.redhat.com/1168405
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit b5d7622f742152d62d850349d7475efed5ff4f3b)