core: do not decrease pending resources twice on run vm
We used to decrease the pending resources twice when running VM - when
the VM is switched to powering-up (onPoweringUp) and when it is switched
to up (runningSucceeded). The problem is that we don't update the cached
field, lastDecreasedVds, so on runningSucceeded we don't know that the
pending resources were already been decreased.

This patch changes the pending decrement in onPoweringUp method such
that the VDS we are decrementing the pending resources from will be
cached.

Change-Id: Ie73f98504ae61ec20a722501fdd49930892dd80a
Signed-off-by: Arik Hadas <ahadas@redhat.com>