core: jobs/steps/audit log handling for domain deactivation
When a Storage Domain is being deactivated, we first attempt to update
the OVF data on the domain and only afterward attempt to deactivate it.
The OVF update phase may contain create volume operations (in case more
OVF stores are needed) and upload image operations afterwards (to update
OVF data contained on the OVF stores).

This patch handle the jobs/steps of the deactivation scenario, so that
on the EXECUTING step all the Upload/Create Image tasks will appear.

Furthermore, when the execution is initiated by the user there's no
need to log that there are tasks, as the tasks were created by the
flow itself (it's being checked on the CDA phase that there are no
tasks prior to the execution).

Change-Id: I4cde77da50fe63a4fda6d9fb61e95fc113e4b54c
Signed-off-by: Liron Aravot <laravot@redhat.com>