core: Set emulated machines after capabilities
After incorrectly adding a ppc64 host in an empty x86-64 cluster and
removing it, the cluster cannot accept new x86-64 hosts, accusing
incompatibility between the emulatedMachines property of the cluster
and the hosts being added.
To fix the issue, the place where emulated machines is set checks if
the host is operational before being executed.

Change-Id: Ibc25acf8e9914e382a5ebf66f9c5d895ed4dab2a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1061320
Signed-off-by: Leonardo Bianconi <leonardo.bianconi@eldorado.org.br>
Signed-off-by: Vitor de Lima <vdelima@redhat.com>