core: Add try/catch block in end action of ExportVmCommand
When exporting a VM with the collapse snapshot option, the VM ovf is
updated after the images export ends successfully. If this operation will
fail and throw an exception the source VM will be stayed in locked status.
This patch surrounds the operation with a try/catch block in order to log
the fault and continue the end action correctly to unlock the VM and it's
disks.

Bug-Url: https://bugzilla.redhat.com/1167297
Change-Id: I67e5119f0f65ce77bc30b6762192e927c88c4434
Signed-off-by: Tal Nisan <tnisan@redhat.com>