restapi: Fix disk move and export metadata
The description of the operations to move, copy and export disks using
the top level disks collection is incorrect. This means that these
operations aren't supported in the SDKs and in the CLI. This patch fixes
the documentation.

Change-Id: I8a5cec5bf3049d3a9325b23a7d54c21565388b51
Bug-Url: https://bugzilla.redhat.com/1169376
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit a3589e9e649f076344b1949aa9cc99944d7de942)