core, engine: Disable CPU pinning validation
This patch disables the validation responsible for checking if a
physical CPU exists during the verification of CPU pinning settings.

This is needed in ppc64 due to the lack of a list of which CPUs are
online, something that is correctly reported currently only in x86-64.

As soon as this missing info is fixed in ppc64, this patch can be reverted.

Change-Id: I98c7a2f136b278a7a695bae8b2eadd3658f3e65c
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1171724
Signed-off-by: Vitor de Lima <vdelima@redhat.com>