core: Update Vm dynamic using VmManager
update the VM dynamic using the VmManager. Its should be quicker and
much less blocking as this is done in a per-vm lock supplied by the
VmManager which means no contention on the Host lock.

also this would help to manage a cached instance of the VM in the
comming future

Change-Id: I65375243d4260b40a0bed1bcccbc48ef83f68a99
Signed-off-by: Roy Golan <rgolan@redhat.com>