core: always perform ovf update after disk creation
Currently the OVF Update is performed after the disk creation only if
atleast one OVF store creation succeeds.

The reason for that is that prior to change I54cc9 the OVF update flow
first updated the existing OVF stores, then created new stores if needed
and if atleast of creation was succesful - update the OVF data.

As in change I54cc9 the order of operations was change (first the OVF
disks are created and then the OVF data is updated) performing the
update only if one disk succeeded is wrong and it should be done always.

Change-Id: Ic029f819f2d0c0a51c098667daa319b921f8750c
Signed-off-by: Liron Aravot <laravot@redhat.com>