aaa: Stop constant synchronization of admin@internal
The following changes are required -
a. when reading group_ids from db - an empty string of group_ids  should not become
a a set containing the "zero" guid, but should be an empty set.
b. group_ids should not have sub strings of "zero" guids
c. equals of DbUser should not check isAdmin

Bug-Url: https://bugzilla.redhat.com/1175046
Change-Id: I5a7660f813b7926b4193ebeb39b3c2a5f07550b4
Topic: AAA
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>