webadmin: Start/stop progress on async operations
Implemented some infrastructure to automatically trigger loading
animation whenever an async operation is invoked, and cancel the
animation whenever the operations returns.

This only happens if upon initialization of the operation, the source
model is "properly" passed. For a single query, the model should be
passed as the target of its AsyncQuery callback. Otherwise (multiple
queries, single/multiple actions), the model should be passed as the
"state" argument.

Also, this only happens in case progress wasn't manually set; if it
had been, then this would override the automatic infrastructure
altogether.

Change-Id: I24ed80a5453668b37f488e0226f7bdfbf36e503d
Bug-Url: https://bugzilla.redhat.com/1167327
Signed-off-by: Lior Vernia <lvernia@redhat.com>