core: hotfix daotests in VmStaticDAOTest
testGetOrderedVmGuidsForRunMultipleActionsByMigrationSupport which
tested db
logic for ordered VmGuids retrieval (with respect to MigrationSupport
attribute) was failing because the number of tested vms was 4, but
MigrationSupport has only 3 possible values. Therefore there were 2 VMs
that
were equal with respect to this ordering.

Hot-fixed by restricting the number of test VMs to 3.

Change-Id: I53df7e6003582ece29387c59312ea9a267f7a1d0
Bug-Url: https://bugzilla.redhat.com/1220116
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>