webadmin: Remove redundant query events
With the addition of the two general async operation events, there's
no need for the existing ones specific to queries - they were all used
to control progress, which is exactly what the new events do (in a
more generic way).

The previous event system relied on computed hash to make sure the
event corresponds to the correct model. This has been swapped for
equality of the model pointer - as the old hash was composed using the
model's hashname and date, this should be equivalent (it is at least
as discriminating).

This also rendered the "context" member of AsyncQuery redundant.

Change-Id: Ia708871f6e4464f894270b19c5a866c50ac7ce8b
Signed-off-by: Lior Vernia <lvernia@redhat.com>