restapi: Add Host Device Passthrough support
Added new endpoints for managing host device passthrough.

The /hosts/{id}/devices endpoint is read-only and reports all devices
that are found on the host.

The /vms/{id}/hostdevices endpoint manages the actual assignment of host
devices to individual VMs.

In all cases the devices are identified with name, but in the API we used
'id' field for this purpose in order to fit properly within the nested resource
and links hierarchy.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1231808
Change-Id: I56a11bf08bb0490ae2a0b481365d3ef38205ff91
Signed-off-by: Martin Betak <mbetak@redhat.com>