core: prevent disks from remaining in locked status
When disks are marked to remain in ILLEGAL status after failed creation
they remain LOCKED because getRelevantDiskImage() retunrs null.
As this patch is about to be backported to the 3.5 branch and it's
preferred to keep the side effects to be minimal, in this patch i
overload the setImageStatus method so that the disk will be passed to
it.

Change-Id: Iaf82438a9985b86b8f5412158f236628972c0f3f
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1172130
Signed-off-by: Liron Aravot <laravot@redhat.com>