core: fix removal of attached disks to stateless vm
Change the way we handle disks when stateless VM goes down, to
prevent situations where disks are being removed. This might happen
when disk is attached to stateless VM while it was running (hot-plug)
so no snapshot was taken for it as part of the stateless snapshot of
the VM, and when restoring the stateless snapshot we'll remove it.

The solution is to find disks which don't have volumes attached to the
stateless snapshot and to detach them from the VM before restoring the
stateless snapshot.

Change-Id: I3a1d6eb8d2f4606622c7ed5c73370792406bb9b3
Bug-Url: https://bugzilla.redhat.com/1161204
Signed-off-by: Arik Hadas <ahadas@redhat.com>