frontend: created instance image widget
Created a new instance images editor which is an AddRemoveRowWidget (similar to
network profiles) .

This editor contains one line per one disk. Each line contains two buttons
(create and attach) for create/attach a disk. If the disk is already defined
for the specific line (either newly after clicking create/attach or the VM
already has this disk), the edit button is shown.

If the edit button is clicked for a line which contains a disk which is not yet
physically created, the same dialog as which was used to define it opens.

When the VM already contains the disk, than on edit Vm the edit disk dialog is
opened after clicking the edit button.

Feature Page: http://www.ovirt.org/Features/Instance_Types

Change-Id: I4bcaa1e40841aa12e64708ec82afdb9daec9fb7f
Bug-Url: https://bugzilla.redhat.com/1117489
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>