aaa: two providers produce the same "domain id"
This may occur if two authz names have the same prefix that gets
converted to the same Guid.
This happens as the Guid ctor that receives an array of bytes needs
only the first 16 bytes to consturct a guid and ignores the rest.
The fix suggest an id that is based on base64 encoding of the entire string.

Change-Id: Ieddb2c31645d7276652db17df190bc4f2cfdaacb
Bug-Url: https://bugzilla.redhat.com/1172187
Topic: AAA
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>