core: refactor; encapsulation
• Map<Guid, Whatever> is only mapping Guid to Whatever. It does not
guarantee, that Guid is ID of Whatever. This should be somehow
strongly enforced; it's not safe passing mappings around forcing
callee to trust caller, that he did mapping right. Touched files was
not interrested in created Maps itself ~ true usecase was: get network
for iface and get network for attachment.

Change-Id: Ic42fb58302a66995f366227e43950d2722c77b33
Signed-off-by: Martin Mucha <mmucha@redhat.com>