core: filter duplicate requests to migrate the same vm
Because of a bug in the UI, we might get several requests together to
migrate the same VM. It produces an error message to the user saying the
VM is already been migrated although it is actually being migrated
because of his request.

The solution is to override the hashCode and equals method in the
parameters of migration so duplicate requests will be filtered by
MultipleActionRunner.

Change-Id: Idda26210b9cf4f2746e27a3291bab20b10aaf103
Signed-off-by: Arik Hadas <ahadas@redhat.com>