engine: Remove compatibility check when updating reported QoS
This should be okay without the check even when the host is too old -
the entry should just not exist, and then be deserialized to null
(which is exactly as avoiding it altogether).

The check had only been put there because of a backwards compatibility
problem with reported QoS between 3.3 and 3.4, but towards 3.6 the
entry name was completely changed, so it can't be reported in a
non-compatible manner anymore.

Change-Id: I4bcb42b9e1a0f5c978c90d02b635ae30570a48aa
Signed-off-by: Lior Vernia <lvernia@redhat.com>