packaging: setup: sharpen the test for engine enabled
Without this change:
1. install and setup engine on machine A
2. install and setup other components (dwh, reports, wsp) on machine B
3. install engine on machine B
4. Run engine-setup - will silently enable the engine

This was done for upgrades from 3.4 that need to enable the engine in
the environment, because in 3.4 it was always enabled and there was no
environment for that.

This change also checks the engine FQDN in engine config files and
compares that with our FQDN. We enable if exists and equal.

Bug-Url: https://bugzilla.redhat.com/1174667
Change-Id: Ie58fcaff0925f3b5ac16fc5159a3eda4498f0b6e
Signed-off-by: Yedidyah Bar David <didi@redhat.com>